requirements.txt

Django==5.2
psycopg2-binary==2.9.10
djangorestframework
whitenoise
gunicorn>=23.0
sqlparse==0.5.0
Pillow==10.3.0
django-taggit==5.0.1
Markdown==3.6
psycopg==3.1.18
python-decouple==3.8
social-auth-app-django==5.4.0
django-extensions==3.2.3
werkzeug==3.0.2
pyOpenSSL==24.1.0
requests==2.31.0
easy-thumbnails==2.8.5
django-debug-toolbar==4.3.0
redis==5.0.4
flower==2.0.1
celery==5.4.0
stripe==9.3.0
WeasyPrint==61.2
django-braces==1.15.0
django-embed-video==1.4.9
pymemcache==4.0.0
django-redisboard==8.4.0
channels[daphne]==4.1.0
channels-redis==4.2.0
uwsgi==2.0.25.1
django-rosetta==0.10.0
django-parler==2.3
django-localflavor==4.0
psutil==6.0.0

